# Performance: Add Identifier helper class