--- a/src/MudBlazor.Docs/Components/SectionContent.razor.cs
+++ b/src/MudBlazor.Docs/Components/SectionContent.razor.cs
@@ -50,7 +50,7 @@ public partial class SectionContent
             .AddClass("show-code", _hasCode && ShowCode)
             .Build();
 
-    private string _snippetId = "_" + Guid.NewGuid().ToString()[..8];
+    private string _snippetId = Identifier.Create();
 
     [Parameter] public string Class { get; set; }
     [Parameter] public bool DarkenBackground { get; set; }