--- a/src/MudBlazor/Components/Picker/MudPicker.razor.cs
+++ b/src/MudBlazor/Components/Picker/MudPicker.razor.cs
@@ -27,7 +27,7 @@ namespace MudBlazor
         [Inject]
         private IKeyInterceptorFactory KeyInterceptorFactory { get; set; }
 
-        private string _elementId = "picker" + Guid.NewGuid().ToString().Substring(0, 8);
+        private string _elementId = Identifier.Create("picker");
 
         protected string PickerClassname =>
             new CssBuilder("mud-picker")