--- a/src/MudBlazor/Utilities/Identifier.cs
+++ b/src/MudBlazor/Utilities/Identifier.cs
@@ -1,4 +1,4 @@
-// Copyright (c) MudBlazor 2021
+﻿// Copyright (c) MudBlazor 2021
 // MudBlazor licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.