Hey,

this PR adds a new helper class `Identifier` which provides a `Create` method to create a unique string with fewer allocations than the current approach.

```c#
// Current
private string _elementId = "checkbox" + Guid.NewGuid().ToString().Substring(0, 8);

// After
private string _elementId = Identifier.Create("checkbox");
```

Benchmark:

| Method            | Mean     | Error    | StdDev   | Ratio | Allocated | Alloc Ratio |
|------------------ |---------:|---------:|---------:|------:|----------:|------------:|
| Guid_SubString    | 85.88 ns | 0.717 ns | 0.635 ns |  1.00 |     192 B |        1.00 |
| Identifier_Create | 78.02 ns | 0.371 ns | 0.347 ns |  0.91 |      56 B |        0.29 |


Currently, I only updated the `MudCheckBox` component. Depending on your feedback, I would update all usages.

## How Has This Been Tested?
Added tests.


## Type of Changes

- [x] New feature (non-breaking change which adds functionality)

## Checklist

- [x] The PR is submitted to the correct branch (`dev`).
- [x] My code follows the code style of this project.
- [x] I've added relevant tests.