@xC0dex 2024-07-24
Hey @henon @ScarletKuro what's your opinion on this idea?