@ScarletKuro 2024-07-25
Wouldn't something like this be better if we chase performance / allocation?

```C#
private const string Chars = "abcdefghijklmnopqrstuvwxyz0123456789";
private const int RandomStringLength = 8;

internal static string Create(ReadOnlySpan<char> prefix)
{
    Span<char> identifierSpan = stackalloc char[prefix.Length + RandomStringLength];
    prefix.CopyTo(identifierSpan);

    Span<byte> randomBytes = stackalloc byte[RandomStringLength];
    RandomNumberGenerator.Fill(randomBytes);

    for (int i = 0; i < RandomStringLength; i++)
    {
        identifierSpan[prefix.Length + i] = Chars[randomBytes[i] % Chars.Length];
    }

    return identifierSpan.ToString();
}
```

I feel like `GUID` is too heavy just to generate 8 chars from a 32-char `GUID`. I don't think we need super uniqueness, and `RandomNumberGenerator` is a true cryptographic random generator.