@henon 2024-07-25
Sure, would suffice. I do something similar in my apps actually.