@xC0dex 2024-07-25
You are right. I didn't think about that, great idea. However, the allocations on the heap are unchanged, but it's faster. Now I have the following code, but "only" the `Random` instance:

```c#
private const string Chars = "abcdefghijklmnopqrstuvwxyz0123456789";
private const int CharsLength = 35;
private const int RandomStringLength = 8;

internal static string Create(ReadOnlySpan<char> prefix)
{
    Span<char> identifierSpan = stackalloc char[prefix.Length + RandomStringLength];
    prefix.CopyTo(identifierSpan);

    for (var i = 0; i < RandomStringLength; i++)
    {
        var index = Random.Shared.Next(CharsLength);
        identifierSpan[prefix.Length + i] = Chars[index];
    }

    return identifierSpan.ToString();
}
```

with these results:

| Method    | Mean     | Error    | StdDev   | Ratio | Allocated | Alloc Ratio |
|---------- |---------:|---------:|---------:|------:|----------:|------------:|
| SubString | 88.83 ns | 1.344 ns | 1.192 ns |  1.00 |     192 B |        1.00 |
| Create    | 66.75 ns | 1.304 ns | 1.220 ns |  0.75 |      56 B |        0.29 |