@xC0dex 2024-07-26
I have three notes:
- I'm not sure about the prefix in `MudComponentBase`. Currently, it's `mudinput` but maybe `mud` or no default prefix would fit better.
- Should the new identifier class be public? Then others could use it as well.
- Are you ok with the name of the class and the method?