@ScarletKuro 2024-07-26
> * I'm not sure about the prefix in `MudComponentBase`. Currently, it's `mudinput` but maybe `mud` or the default prefix would fit better.

As long as it doesn't break any unit tests, etc., then I'm okay with the rename. The shorter, the better.

> * Should the new identifier class be public? Then others could use it as well.

No.

> * Are you ok with the name of the class and the method?

Yes.