@xC0dex 2024-07-26
>As long as it doesn't break any unit tests

Okay, they do break. I'll leave it as it is. PR is ready 👍.